chore: enhance base dataset for loong (#1837)
