+++ camel/datasets/__init__.py
@@ -11,14 +11,16 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ========= Copyright 2023-2024 @ CAMEL-AI.org. All Rights Reserved. =========
-from .base import (
-    DataPoint,
-    GenerativeDataset,
-    StaticDataset,
+from .base_generator import (
+    BaseGenerator,
+    FewShotGenerator,
 )
+from .models import DataPoint
+from .static_dataset import StaticDataset
 
 __all__ = [
+    "BaseGenerator",
     "DataPoint",
+    "FewShotGenerator",
     "StaticDataset",
-    "GenerativeDataset",
 ]
